Run the `Encoding::Converter` test
Merge test files for easier handling.
